+++ @navikt/core/react/src/data/stories/Data.stories.tsx
@@ -1,20 +1,28 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import {
-  ColumnPinningState,
+  Table,
   flexRender,
   getCoreRowModel,
   getFilteredRowModel,
+  getPaginationRowModel,
+  getSortedRowModel,
   useReactTable,
 } from "@tanstack/react-table";
-import React, { useState } from "react";
-import { CogIcon, RectangleSectionsIcon } from "@navikt/aksel-icons";
+import React from "react";
+import {
+  CheckmarkIcon,
+  CogIcon,
+  Density1Icon,
+  RectangleSectionsIcon,
+} from "@navikt/aksel-icons";
 import { Button } from "../../button";
 import { VStack } from "../../layout/stack";
 import { ActionMenu } from "../../overlays/action-menu";
 import DataActionBar from "../action-bar/root/DataActionBarRoot";
 import { DataTable } from "../table";
 import { DataToolbar } from "../toolbar";
-import { columns, sampleData } from "./dummy-data";
+import { DataTableProfiler } from "./DataTableProfiler";
+import { PersonInfo, columns, sampleData } from "./dummy-data";
 
 const meta: Meta<typeof DataTable> = {
   title: "ds-react/Data",
@@ -23,6 +31,7 @@ const meta: Meta<typeof DataTable> = {
     chromatic: { disable: true },
     layout: "padded",
   },
+  decorators: [(Story) => <DataTableProfiler>{Story()}</DataTableProfiler>],
 };
 
 export default meta;
@@ -74,34 +83,48 @@ export const Default: Story = {
 
 export const TanstackExample: Story = {
   render: () => {
-    const [globalFilter, setGlobalFilter] = useState<string>();
-    const [columnVisibility, setColumnVisibility] = React.useState({});
-    const [columnPinning, setColumnPinning] =
-      React.useState<ColumnPinningState>({ left: [], right: [] });
-
+    const [rowDensity, setRowDensity] = React.useState<
+      "normal" | "condensed" | "spacious"
+    >("normal");
     const table = useReactTable({
       columns,
       data: sampleData,
       getCoreRowModel: getCoreRowModel(),
+      getSortedRowModel: getSortedRowModel(),
       getFilteredRowModel: getFilteredRowModel(),
       globalFilterFn: "includesString",
-      state: {
-        globalFilter,
-        columnVisibility,
-        columnPinning,
+      getPaginationRowModel: getPaginationRowModel(),
+      initialState: {
+        pagination: {
+          pageIndex: 1,
+          pageSize: 20,
+        },
       },
-      onColumnPinningChange: setColumnPinning,
-      onGlobalFilterChange: setGlobalFilter,
-      onColumnVisibilityChange: setColumnVisibility,
+      state: {},
       columnResizeMode: "onChange",
+      debugTable: true,
+      debugHeaders: true,
+      debugColumns: true,
     });
+
+    const columnSizeVars = () => {
+      const headers = table.getFlatHeaders();
+      const colSizes: { [key: string]: `${number}px` } = {};
+      for (let i = 0; i < headers.length; i++) {
+        const header = headers[i];
+        colSizes[`--header-${header.id}-size`] = `${header.getSize()}px`;
+        colSizes[`--col-${header.column.id}-size`] =
+          `${header.column.getSize()}px`;
+      }
+      return colSizes;
+    };
 
     return (
       <VStack gap="space-16">
         <DataToolbar>
           <DataToolbar.SearchField
             label="Tekstfilter"
-            onChange={(e) => setGlobalFilter(e)}
+            onChange={(value) => table.setGlobalFilter(value)}
           />
           <DataToolbar.ToggleButton icon={<RectangleSectionsIcon />} />
           <ActionMenu>
@@ -139,6 +162,42 @@ export const TanstackExample: Story = {
               })}
             </ActionMenu.Content>
           </ActionMenu>
+          <ActionMenu>
+            <ActionMenu.Trigger>
+              <Button
+                data-color="neutral"
+                variant="tertiary"
+                size="small"
+                icon={<Density1Icon title="Tetthet" />}
+              />
+            </ActionMenu.Trigger>
+            <ActionMenu.Content>
+              <ActionMenu.Group aria-label="Velg tetthet">
+                <ActionMenu.Item
+                  onSelect={() => setRowDensity("condensed")}
+                  icon={
+                    rowDensity === "condensed" ? <CheckmarkIcon /> : undefined
+                  }
+                >
+                  Tett
+                </ActionMenu.Item>
+                <ActionMenu.Item
+                  onSelect={() => setRowDensity("normal")}
+                  icon={rowDensity === "normal" ? <CheckmarkIcon /> : undefined}
+                >
+                  Normal
+                </ActionMenu.Item>
+                <ActionMenu.Item
+                  onSelect={() => setRowDensity("spacious")}
+                  icon={
+                    rowDensity === "spacious" ? <CheckmarkIcon /> : undefined
+                  }
+                >
+                  Løs
+                </ActionMenu.Item>
+              </ActionMenu.Group>
+            </ActionMenu.Content>
+          </ActionMenu>
         </DataToolbar>
 
         <DataActionBar numOfSelectedRows={2} onClear={() => alert("Cleared!")}>
@@ -150,7 +209,7 @@ export const TanstackExample: Story = {
           </Button>
         </DataActionBar>
 
-        <DataTable style={{ width: "3000px" }}>
+        <DataTable style={columnSizeVars()} rowDensity={rowDensity}>
           <DataTable.Thead>
             {table.getHeaderGroups().map((headerGroup) => {
               return (
@@ -159,14 +218,20 @@ export const TanstackExample: Story = {
                     return (
                       <DataTable.Th
                         key={header.id}
-                        size={header.getSize()}
+                        style={{ width: `var(--header-${header.id}-size)` }}
                         resizeHandler={header.getResizeHandler()}
-                        pinningHandler={
+                        /* pinningHandler={
                           header.column.getIsPinned() === "left"
                             ? () => header.column.pin(false)
                             : () => header.column.pin("left")
                         }
-                        isPinned={header.column.getIsPinned() === "left"}
+                        isPinned={header.column.getIsPinned() === "left"} */
+                        sortDirection={header.column.getIsSorted() || "none"}
+                        onSortChange={(_, event) => {
+                          const handler =
+                            header.column.getToggleSortingHandler();
+                          handler?.(event);
+                        }}
                       >
                         {header.isPlaceholder
                           ? null
@@ -181,26 +246,61 @@ export const TanstackExample: Story = {
               );
             })}
           </DataTable.Thead>
-          <DataTable.Tbody>
-            {table.getRowModel().rows.map((row) => {
-              return (
-                <DataTable.Tr key={row.id}>
-                  {row.getVisibleCells().map((cell) => {
-                    return (
-                      <DataTable.Td key={cell.id}>
-                        {flexRender(
-                          cell.column.columnDef.cell,
-                          cell.getContext(),
+
+          {table.getState().columnSizingInfo.isResizingColumn ? (
+            <MemoizedTableBody table={table} />
+          ) : (
+            <TableBody table={table} />
+          )}
+          <DataTable.Tfoot>
+            {table.getFooterGroups().map((footerGroup) => (
+              <DataTable.Tr key={footerGroup.id}>
+                {footerGroup.headers.map((header) => (
+                  <DataTable.Td
+                    key={header.id}
+                    style={{ width: `var(--header-${header.id}-size)` }}
+                  >
+                    {header.isPlaceholder
+                      ? null
+                      : flexRender(
+                          header.column.columnDef.footer,
+                          header.getContext(),
                         )}
-                      </DataTable.Td>
-                    );
-                  })}
-                </DataTable.Tr>
-              );
-            })}
-          </DataTable.Tbody>
+                  </DataTable.Td>
+                ))}
+              </DataTable.Tr>
+            ))}
+          </DataTable.Tfoot>
         </DataTable>
       </VStack>
     );
   },
+  parameters: {
+    a11y: { disable: true },
+    controls: { disable: true },
+    docs: { disable: true },
+  },
 };
+
+const TableBody = ({ table }: { table: Table<PersonInfo> }) => (
+  <DataTable.Tbody>
+    {table.getRowModel().rows.map((row) => {
+      return (
+        <DataTable.Tr key={row.id}>
+          {row.getVisibleCells().map((cell) => {
+            return (
+              <DataTable.Td key={cell.id}>
+                {flexRender(cell.column.columnDef.cell, cell.getContext())}
+              </DataTable.Td>
+            );
+          })}
+        </DataTable.Tr>
+      );
+    })}
+  </DataTable.Tbody>
+);
+
+const MemoizedTableBody = React.memo(
+  TableBody,
+  (prev, next) => prev.table.options.data === next.table.options.data,
+) as typeof TableBody;