+++ aksel.nav.no/website/proxy.ts
@@ -1,6 +1,7 @@
 import type { NextRequest } from "next/server";
 import { NextResponse } from "next/server";
-import { noCdnClient, sanityClient } from "@/sanity/client.server";
+import { AVALIABLE_MARKDOWN_ROUTES } from "@/app/api/markdown/llm.config";
+import { sanityClient } from "@/sanity/client.server";
 
 const ignoredPaths = ["/eksempler", "/templates", "/ikoner", "/admin"];
 const ignoredStaticPaths = [
@@ -27,6 +28,10 @@ export async function proxy(req: NextRequest) {
     return NextResponse.next();
   }
 
+  const { pathname } = req.nextUrl;
+
+  const url = req.nextUrl.clone();
+
   if (
     ignoredPaths.some((prefix) => req.nextUrl.pathname.startsWith(prefix)) ||
     ignoredStaticPaths.some((prefix) => req.nextUrl.pathname === prefix)
@@ -34,7 +39,6 @@ export async function proxy(req: NextRequest) {
     return NextResponse.next();
   }
 
-  const url = req.nextUrl.clone();
   if (
     url.pathname.startsWith("/sandbox") &&
     !url.pathname.includes("index.html") &&
@@ -45,6 +49,9 @@ export async function proxy(req: NextRequest) {
   }
 
   try {
+    /**
+     * TODO: Look into updating this using tag-based revalidation
+     */
     const redirect = await sanityClient.fetch(
       `
   *[_type == 'redirect' && source == $source][0] {
@@ -54,21 +61,36 @@ export async function proxy(req: NextRequest) {
   }
 `,
       { source: decodeURIComponent(req.nextUrl.pathname) },
+      { cache: "force-cache", next: { revalidate: 3600 } },
     );
 
     if (redirect) {
-      const token = process.env.SANITY_WRITE;
+      /**
+       * TODO: Temp disabled due to revalidation issues causing excessive re-validations
+       */
+      /* const token = process.env.SANITY_WRITE;
       if (token) {
         noCdnClient(token)
           .patch(redirect._id)
           .set({ redirects: 1 + (redirect.redirects ?? 0) })
           .commit();
-      }
+      } */
 
       if (redirect.destination.startsWith("http")) {
         return NextResponse.redirect(new URL(redirect.destination));
       }
       return NextResponse.redirect(new URL(redirect.destination, req.url));
+    }
+
+    /* Check if the request is for a markdown version (.md extension) */
+    if (
+      pathname.endsWith(".md") &&
+      AVALIABLE_MARKDOWN_ROUTES.includes(pathname.slice(0, -3))
+    ) {
+      /* Rewrite to the markdown API route with the original path as a parameter */
+      url.pathname = "/api/markdown";
+
+      return NextResponse.rewrite(url);
     }
 
     return NextResponse.next();