+++ @navikt/core/react/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @navikt/ds-react
 
+## 8.3.0
+
+### Minor Changes
+
+- Tooltip: Updated prop 'keys' to allow for multiple shortcuts. ([#4604](https://github.com/navikt/aksel/pull/4604))
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @navikt/ds-tokens@8.3.0
+  - @navikt/aksel-icons@8.3.0
+
+## 8.2.2
+
+### Patch Changes
+
+- ExpansionCard: Remove redundant class name ([#4602](https://github.com/navikt/aksel/pull/4602))
+
+- AccordionItem: Remove redundant class name ([#4602](https://github.com/navikt/aksel/pull/4602))
+
+- Updated dependencies []:
+  - @navikt/ds-tokens@8.2.2
+  - @navikt/aksel-icons@8.2.2
+
+## 8.2.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @navikt/ds-tokens@8.2.1
+  - @navikt/aksel-icons@8.2.1
+
 ## 8.2.0
 
 ### Minor Changes